api: Extracted monitoring to separate file
